moving junit testing module to htmresearch/support
